data-engine-api/pyproject.toml
[project]
name = "data-engine-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-cors>=4.0.0",
    "flasgger>=0.9.7.1",
    "langchain>=1.0.5",
    "langchain-core>=0.3.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
    "psycopg2-binary>=2.9.9"
    ]
